backend/requirements.txt
Django>=2.0,<3.0
djangorestframework>=3.0,<4.0
gunicorn>=19.0,<20.0
psycopg2-binary>=2.0,<3.0
